+++ ZapLibUnitTest/ImplicitMailerTests.cs
@@ -23,7 +23,7 @@ namespace ZapLib.Tests
 
             string SMTP_HOST = "smtp.gmail.com";
             string MAIL_ACT = "maryzap123@gmail.com";
-            string MAIL_PWD = "";
+            string MAIL_PWD = "mwydfkirmrtfvxql";
 
 
             int MAIL_PORT = 465;
@@ -31,12 +31,14 @@ namespace ZapLib.Tests
 
             ImplicitMailer m = new ImplicitMailer(SMTP_HOST, MAIL_ACT, MAIL_PWD, MAIL_PORT, MAIL_SSL);
 
-            string TO = "zaplin@iqs-t.com,sollin@iqs-t.com";
+            string TO = "zaplin@iqs-t.com,sollin@iqs-t.com,brucewang@iqs-t.com";
             string SUBJECT = $"This is a test mail (ZapLib.ImplicitMailer) - {DateTime.Now}";
             string BODY = $"<h1>Test Mail by ZapLib ImplicitMailer</h1>";
             string[] Attachment = new string[] { @"D:\Downloads\delivered.png", @"D:\Downloads\abstract-christmas-tree.zip", @"D:\Downloads\.gitignore" };
-
-
+            string cid = m.AddAttachments(@"D:\Downloads\user.png");
+            foreach(var a in m.AttachmentsList) {
+                BODY += $"<p>圖片在這: <img src=\"cid:{cid}\" ></p>";
+            }
             bool result = m.Send(TO, SUBJECT, BODY, attchments: Attachment);
             Trace.WriteLine(m.ErrMsg);
             Assert.IsTrue(result);